✨ Configure axios retry with backoff
The node sdk currently user axios-retry, but provides no default configuration. This configures it to retry up to 3 times on failure
with an exponential delay between these, to help guard against constantly bumping into our rate-limits.

Needed to upgrade a few dependencies to get this working and the builds running.

See VI-2000