♻️ Standardize application client id
Currently to initialize the API client your need to provide the application
client_id in the optional data but this isn't actually optional as we throw
an error if it isn't provided. It would be much clearer for developers if
this was just a field of the constructor.

See Issue#129