🐛 Rewrite search query builder for non-accounting resources
Currently all the query parameters generated by SearchQueryBuilder are
in the form expected by FreshBooks accounting resources. However other
resource expect different forms, in particular for `equals` parameters.

This changes the builder to store data on the query at the time it is
declared and generate it at the time of use as we don't know the type of
resource until the call is made.

Fixes #235